refactor(*): rename mocks folder to __mocks__
ref: #MANAGER-17422
