+++ packages/manager-react-components/src/__mocks__/images.tsx
