+++ packages/manager-react-components/src/__mocks__/stepper.ts
