+++ packages/manager-react-components/src/__mocks__/tiles-input.ts
