+++ packages/manager/apps/hub/src/__mocks__/catalog.ts
