+++ packages/manager/apps/hub/src/__mocks__/products.ts
