+++ packages/manager/apps/pci-rancher/src/__mocks__/catalog.ts
