+++ packages/manager/apps/pci-rancher/src/__mocks__/images.tsx
