+++ packages/manager/apps/pci-rancher/src/__mocks__/rancher-resource.ts
