+++ packages/manager/apps/pci-rancher/src/__mocks__/version.ts
