+++ packages/manager/apps/pci-rancher/src/components/Modal/DeleteModal/DeleteModal.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event';
 import listingTranslation from '@translation/listing/Messages_fr_FR.json';
 import DeleteModal, { TERMINATE_TEXT } from './DeleteModal.component';
 import { fireEvent, render, waitFor, act } from '@/utils/test/test.provider';
-import { rancherMocked } from '@/_mock_/rancher';
+import { rancherMocked } from '@/__mocks__/rancher';
 
 const onDeleteMocked = vi.fn();
