+++ packages/manager/apps/pci-rancher/src/components/Modal/UpdateSoftwareConfirmModal/UpdateSoftwareConfirmModal.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import React from 'react';
 import updateTranslation from '@translation/updateSoftware/Messages_fr_FR.json';
 import dashboardTranslation from '@translation/dashboard/Messages_fr_FR.json';
-import { versionsMocked } from '@/_mock_/version';
+import { versionsMocked } from '@/__mocks__/version';
 import { render, waitFor } from '@/utils/test/test.provider';
 import UpdateSoftwareModal from './UpdateSoftwareConfirmModal.component';
