+++ packages/manager/apps/pci-rancher/src/components/Table/ActionsCell/ActionsCell.test.tsx
@@ -3,7 +3,7 @@ import { Row } from '@tanstack/react-table';
 import listingTranslation from '@translation/listing/Messages_fr_FR.json';
 import ActionsCell from './ActionsCell.component';
 import { render, waitFor } from '@/utils/test/test.provider';
-import { rancherError, rancherMocked } from '@/_mock_/rancher';
+import { rancherError, rancherMocked } from '@/__mocks__/rancher';
 import { RancherService } from '@/types/api.type';
 
 const setupSpecTest = async (rancherService: RancherService = rancherMocked) =>
