+++ packages/manager/apps/pci-rancher/src/components/UpdateRancherVersionBanner/UpdateVersionBanner.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import updateSoftwareTranslation from '@translation/updateSoftware/Messages_fr_FR.json';
 import { render, waitFor } from '@/utils/test/test.provider';
-import { rancherMocked } from '@/_mock_/rancher';
-import { versionsMocked } from '@/_mock_/version';
+import { rancherMocked } from '@/__mocks__/rancher';
+import { versionsMocked } from '@/__mocks__/version';
 import { ResourceStatus } from '@/types/api.type';
 import UpdateVersionBanner from './UpdateVersionBanner.component';
