+++ packages/manager/apps/pci-savings-plan/src/__mocks__/commercial-catalog/technical.ts
