+++ packages/manager/modules/order/__mocks__/order-details.ts
